[MIRROR] [MDB IGNORE] RTD, RLD and RSF Prettification (#5054)
Original PR: https://github.com/tgstation/tgstation/pull/74093
-----

## About The Pull Request

Remade the RTD and its charge meter in the style of the other rapid
tools using the RCD as a base, along with its in hand icons.
Altered the RLD's in hand icons to more closely match their inventory
icon's palette and appearance.
Altered the RSF's darkest green color in its palette to be actually
green.
## Why It's Good For The Game
I think updating the RTD speaks for itself. The old one is just ugly and
sticks out like a sore thumb amongst the rest of the tools and honestly
seems like a placeholder.
the RLD's hand icons didn't match its inventory icon in its palette and
details. now it does and looks more consistent.
the last change on the RSF might be minutiae. if you cant tell what it
is: the very dark orange from the orange tools was kept despite the
green palette swap. which was most likely a mistake. so I gave it a dark
green that goes well with the rest of the palette even if it was only
for a few pixels.


![image](https://user-images.githubusercontent.com/121331795/226132541-d99d5723-dfe6-4ffd-a861-6d997d090a25.png)
(The RLD inventory icon is just there for reference.)
## Changelog
:cl:

imageadd: prettified all RTD icons 
imageadd: prettified in hand RLD icons 
imageadd: subtly improved the RSF icons

/:cl:

---------

Co-authored-by: Chlorotrifluoride <121331795+Chlorotrifluoride@users.noreply.github.com>