[MIRROR] [MDB IGNORE] Hermit Ruin Active Turf Fix(?) (#5129)
Original PR: https://github.com/tgstation/tgstation/pull/74306
-----
Note for the future: The hypothesis presented within was later found
that mobs breaking down the walls couldn't cause the active turfs.
However, there is something sus about how thin the walls are and how we
keep observing active turfs along those aspects (all fairy grass is
placed next to the snow walls). One portion of this PR is including more
tools for tracking this ruin, so we'll know if it starts to be a problem
as well.

If it gets to be really bad I'll start including stuff like the turf's
contents in the post-mortem report.
## About The Pull Request

On Round 202559:

```txt
[2023-03-27 07:25:21.791]
 - All that follows is a turf with an active air difference at roundstart. To clear this, make sure that all of the turfs listed below are connected to a turf with the same air contents.
 - In an ideal world, this list should have enough information to help you locate the active turf(s) in question. Unfortunately, this might not be an ideal world.
 - If the round is still ongoing, you can use the "Mapping -> Show roundstart AT list" verb to see exactly what active turfs were detected. Otherwise, good luck.
 - Active turf: Icemoon Underground (174,130,2) (/area/icemoon/underground/explored). Turf type: /turf/open/misc/asteroid/snow/icemoon. Relevant Z-Trait(s): Mining, Ice Ruins Underground and Station.
 - Active turf: Shuttle (175,130,2) (/area/ruin/powered/shuttle). Turf type: /turf/open/floor/grass/fairy. Relevant Z-Trait(s): Mining, Ice Ruins Underground and Station.
 - Active turf: Icemoon Underground (175,131,2) (/area/icemoon/underground/explored). Turf type: /turf/open/misc/asteroid/snow/icemoon. Relevant Z-Trait(s): Mining, Ice Ruins Underground and Station.
 - Active turf: Shuttle (176,131,2) (/area/ruin/powered/shuttle). Turf type: /turf/open/floor/grass/fairy. Relevant Z-Trait(s): Mining, Ice Ruins Underground and Station.
 - Z-Level 2 has 4 active turf(s).
 - Z-Level trait Ice Ruins Underground has 4 active turf(s).
 - Z-Level trait Mining has 4 active turf(s).
 - Z-Level trait Station has 4 active turf(s).
 - End of active turf list.
```

I have no idea what caused it, this is my best guess:


![image](https://user-images.githubusercontent.com/34697715/228062120-cdd414e5-6a2d-4d2f-95ac-921ddcb06d59.png)

Maybe a mob could have broken it before air loaded? unsure, this is a
post-mortem examination so i don't know exactly what failed. I ran
icebox with this ruin as `always_place` at least five times but never
saw any active turfs. Out of an abundance of caution let's cut down how
"thin" the walls are because the active turf could have been spontaneous
from some ruinloader mishap? It's a bit confusing and I don't have a
definite answer on what causes this, so will have to keep a closer eye
on this.

I also added a new ruin area subtype tailored for the hermit's ruin so
we can immediately detect if this should happen again, since I have only
the one data point. i also fixed up some other placement stuff while in
the area.
## Why It's Good For The Game

Hopefully cut down on the number of active turfs we get on production,
since mobs breaking turfs that they can easily break leading to a
failure state is very silly.
## Changelog
:cl:
fix: The hermit of the Icemoon has decided to build their hut in some
deeper caves for a bit more protection from the local fauna.
/:cl:

---------

Co-authored-by: san7890 <the@san7890.com>