[MIRROR] [MDB IGNORE] Fixes suffocation damage and healing for reagents (#5139)
Original PR: https://github.com/tgstation/tgstation/pull/74292
-----

## About The Pull Request

This fixes a hidden side effect of the "refactor" in #73147 that
actually caused ALL existing suffocation damage dealing/healing
chemicals to not work on species that breathe gases other than oxygen.

This is also good for future code since if you want a chemical to not
affect a specific species or such, you should explicitly define that, it
shouldn't be implicitly done for you in the background with 0
implication.

Also reverts #74247 since it's no longer needed.
## Why It's Good For The Game

First off, the change was unintended. It was intended to be a refactor.

Second, the change is unfun and isn't explained anywhere. It isn't very
fun to sit on the floor in medbay while even the most experienced
doctors can't figure out why the strongest suffocation healing chemicals
in the game aren't working. Cue this on repeat for 30 minutes. I've
personally experienced this like 10 times now and even I didn't know why
it was happening.

Even if you know the change exists, it's still extremely convoluted to
get around. You have to do CPR on a plasmaman at the speed of an average
cheetah or abuse the brute/burn revival method.
## Changelog
:cl:
fix: Suffocation-handling reagents work on all species once more.
spellcheck: Added a previous tip for handling plasmaman suffocation back
into the game.
/:cl:

---------

Co-authored-by: RikuTheKiller <88713943+RikuTheKiller@users.noreply.github.com>