Fugitive backstories are now selected with consideration to the number of signups (#74347)
## About The Pull Request

Adjusts how the fugitives event decides to select its backstories.

Formerly, the game would decide if it was polling for a solo (Waldo) or
team (4 player) backstory first, meaning the event could fail if only 3
people signed up when the event decided to run a team backstory. It also
meant that, even with thirty people signed up, a solo backstory (still
just Waldo for now) could be selected. What a waste!

Now, the event receives its signups and uses them to decide which
backstory to choose. If there aren't enough players for a team, a solo
backstory is chosen. If a team can be formed, the event may still decide
to run a solo backstory, but the chance scales inversely with the number
of candidates.

This also adds an atmos safety check to their spawn location finder,
because I was in the area.
## Why It's Good For The Game

Helps a deadchat sink even be more of a deadchat sink. Waldo is a
bastard.
## Changelog
:cl:
balance: Fugitive team backstories will now be selected with
consideration to the number of people who sign up.
qol: Fugitives can no longer spawn in atmos-unsafe areas.
/:cl: