[MIRROR] [MDB IGNORE] Fixes Active Turfs on WayStation (#5159)
Original PR: https://github.com/tgstation/tgstation/pull/74378
-----

## About The Pull Request


![image](https://user-images.githubusercontent.com/34697715/228984205-8d431ae7-fbbb-4a36-b46f-fb81ec9973ac.png)

AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA

This ruin caused 37 out of the average 41 active turfs that you see
above, stay tuned for a PR to fix those too.

Basically we used catwalk floor _turfs_ here instead of the catwalk
structure:


![image](https://user-images.githubusercontent.com/34697715/228984311-6724972c-4240-4d4e-a463-3e1f541b49b6.png)

Simple issue to make, and to fix. We also used space turfs here instead
of the `template_noop` turfs that we should be using for ruins like
this, so that was fixed as well.
## Why It's Good For The Game

Less active turfs. I NEED LESS NOISE SO I CAN FIGURE OUT THE EDGE CASES
SO WE CAN FINALLY MAKE THIS A CI STEP
## Changelog
doesn't matter as far as players go

---------

Co-authored-by: san7890 <the@san7890.com>