[MIRROR] [MDB IGNORE] Fixes "Atmos Asteroid" Active Turfs - Irony Edition (#5162)
Original PR: https://github.com/tgstation/tgstation/pull/74379
-----

## About The Pull Request

This ruin was consistently triggering 4 Active Turfs on prod. 


![image](https://user-images.githubusercontent.com/34697715/228991150-e90e7d8a-a19a-41ff-8101-a3c172e2883a.png)

This was because two turfs didn't have the correct gas mix varedit. This
is probably fine, but I hate varediting turfs like this because anyone
can grab a turf and not realize that it's somehow special if they don't
check the varedits/keys of the map after it's applied. I get the gimmick
and that's cool, but let's do it a different way.

So, I just made them all defined in the code, and span up a quick
updatepaths to update all the turfs (to ensure I didn't balls it up
somehow). Very cool.

I also did two more things

* Added a new area for this ruin only. This is because the atmos turf
stuff is a bit spooky to me, and I want to know _instantly_ if this ruin
is causing more issues. Don't think it should have a name for GPS
reasons but we definitely need a unique area type.

* Fixed a typo in the datum typepath. That was irritating me. It's fixed
now.
## Why It's Good For The Game

I NEED LESS ACTIVE TURFS THE NOISE IS TOO MUCH THE NOISE IS TOO MUCH THE
NOISE IS TOO MUCH THE NOISE IS TOO MUCH THE NOISE IS TOO MUCH THE NOISE
IS TOO MUCH THE NOISE IS TOO MUCH THE NOISE IS TOO MUCH THE NOISE IS TOO
MUCH THE NOISE IS TOO MUCH THE NOISE IS TOO MUCH THE NOISE IS TOO MUCH
THE NOISE IS TOO MUCH THE NOISE IS TOO MUCH THE NOISE IS TOO MUCH THE
NOISE IS TOO MUCH THE NOISE IS TOO MUCH THE NOISE IS TOO MUCH THE NOISE
IS TOO MUCH THE NOISE IS TOO MUCH THE NOISE IS TOO MUCH THE NOISE IS TOO
MUCH THE NOISE IS TOO MUCH THE NOISE IS TOO MUCH
## Changelog
Player's shouldn't really notice or care.

---------

Co-authored-by: san7890 <the@san7890.com>