OOC-ifies the Bluespace Tag (#74203)
## About The Pull Request
De-ICs the bluespace tag, instead presenting the same information in an
OOC format.

![image](https://user-images.githubusercontent.com/58124831/227378515-ccb9bfb3-bd30-4ceb-8c6c-fdc50132fb4b.png)
This does not alter the actual information shown (besides a slight
change to wording for the armor-piercing/block values, though the
thresholds are the same).

P.S. I don't know if this is tagged correctly- it doesn't seem to neatly
fit any category. (When are we getting the Tweak tag back?)
## Why It's Good For The Game
The bluespace tag has sorta been an oddity since it was introduced.
While the rationale behind its addition was sound- we want players to
have information without code-diving, but not necessarily anything too
concrete- the actual stuff surrounding that information has always felt
off. Why exactly are we finding items that have lain undisturbed on
lavaland for centuries that have a bluespace tag? What magical source of
information are they drawing from? Basically, in its attempt to make OOC
information IC-relevant, it creates more plot holes than it solves, and
ultimately it does so to no real benefit. Hence, I feel it's best that
we just stop trying to present this in an IC format.
## Changelog
:cl:
qol: Bluespace tags are gone. The information contained within is still
available in the same way, just in a more out-of-character format.
/:cl: