[MIRROR] [MDB IGNORE] Icemoon Hermit Ruin Active Turf Fix - For Real This Time (#5207)
Original PR: https://github.com/tgstation/tgstation/pull/74476
-----
## About The Pull Request

In #74306, I _thought_ I knew what the cause was, and I both attempted a
potential fix _and_ made tracking it easier. The fruits of my labor paid
off, I know exactly what caused it now.


![image](https://user-images.githubusercontent.com/34697715/229420637-ff4ff886-fa34-4962-8019-d3105bf3d3c3.png)

Basically, the demonic portal will scrape away all turfs in a 5-tile
radius on its `Initialize()`, and if a spawner spawned right next to the
hermit ruin... it would count it as a mineral turf and scrape it away as
well. That's so fucking silly. At least we know now.
## Why It's Good For The Game

The fix is to just make those tiles unscrapeable, which is accomplished
via another turf_flag and filtering those out in the `Initialize()` of
the demonic portals.

I also cleaned up the calls to scrapeaway being `null`, which is really
weird because it just defaulted to the normal proc behavior. Naming the
arguments instead does the same thing (I checked):


![image](https://user-images.githubusercontent.com/34697715/229421028-ffe9c2dd-eaec-4c71-ac8e-96d3a7fa9a0f.png)

The hermit's shelter is finally intact.
## Changelog
Nothing that really concerns players.

---------

Co-authored-by: san7890 <the@san7890.com>
Co-authored-by: Jolly-66 <70232195+Jolly-66@users.noreply.github.com>