[MIRROR] [MDB IGNORE] [NO GBP] Fix minigames UI not removing inactive clients (#5252)
Original PR: https://github.com/tgstation/tgstation/pull/74391
-----
## About The Pull Request
Fixes #74260

The UI for the basketball menu was not properly updating when clients
would disconnect. This would allow someone to signup twice leading to
silly situations as detailed in the issue report. This issue also
affected the mafia minigame due to both UIs having similar code.

The fix is simple, just check the signups when the tgui menu has
interactions and also check to make sure the signups use a boolean
instead of a client for the value part of the key/value list. Why? After
a client would disconnect, the list would change and remove the client
object since it no longer exists. The solution is to keep the ckey as a
key, but use a boolean for the value.

## Why It's Good For The Game
Less bugs, more stability.

## Changelog
:cl:
fix: Fix basketball and mafia minigame UI not removing inactive clients
/:cl:

---------

Co-authored-by: Tim <timothymtorres@gmail.com>
Co-authored-by: Fikou <23585223+Fikou@users.noreply.github.com>