[MIRROR] [MDB IGNORE] Adjustable Goliath Cloaks (#5362)
Original PR: https://github.com/tgstation/tgstation/pull/74709
-----
## About The Pull Request
I've enabled goliath cloaks to be alt clicked to adjust it to neck slot
instead of over clothing slot.
While adjusted to neck slot, it has no armor values for balance
purposes, while in over clothing slot, it has the same armor values as
usual. The hood has the same values as usual, as it still inhabits the
same slot as usual.


![image](https://user-images.githubusercontent.com/126404225/231837773-97f79c0e-9de2-4bfa-bfec-ed6c67e0d476.png)

![image](https://user-images.githubusercontent.com/126404225/231837688-f1f6ea67-8127-423e-96a0-b125f7778194.png)

![image](https://user-images.githubusercontent.com/126404225/231837716-805f778b-4db1-4dd5-a667-c196524f023f.png)
## Why It's Good For The Game
It enables a cloak to function like a cloak. It gives people the option
to use it not for armor, but for fashion.
## Changelog
:cl:
add: Goliath cloak can be worn as a cloak that doesn't provide any armor
benefits
/:cl:

---------

Co-authored-by: Dawnseer <126404225+Dawnseer@users.noreply.github.com>